got function operator in a decent state
